Updated "magic function properties"
The last example didn't actually work as `f` was still set either to the first example (`f = function() {}`), or it was assumed that `f` was supposed to be part of `myFunction`. In either case, it didn't work. Changed `var myFunction` to `f` which also makes it less confusing as to where `.name` gets it's name from.